Spliting out resources and context header files
