+++ include/camp/resource/platform.hpp
@@ -1,0 +1,23 @@
+#ifndef __CAMP_PLATFORM_HPP
+#define __CAMP_PLATFORM_HPP
+
+namespace camp
+{
+namespace resources
+{
+  inline namespace v1
+  {
+
+    enum class Platform {
+      undefined = 0,
+      host = 1,
+      omp = 2,
+      tbb = 4,
+      cuda = 8,
+      hip = 16
+    };
+
+  }  // namespace v1
+}  // namespace resources
+}  // namespace camp
+#endif /* __CAMP_PLATFORM_HPP */
